feat(reducer): split reducer logic
